refactor: consolidate duplicate formatting utilities and shared constants
- Rename getApColor → getAccessPackageColor in utils/colors.js
- Add formatDateOnly, formatDurationSeconds, formatDurationMs,
  formatRelativeTime, formatCompactNumber to utils/formatters.js
- Extract ASSIGNMENT_TYPE_STYLES to utils/accessPackageStyles.js;
  AccessPackagesPage gains dark mode badge classes it was missing
- Remove inline duplicates from SyncLogPage, CrawlersPage, RunDetailPage,
  DashboardPage, AccessPackagesPage, AccessPackageDetailPage, excelHelpers
