+++ app/ui/src/components/AccessPackageDetailPage.jsx
@@ -11,18 +11,12 @@ import RecentChangesSection from './RecentChangesSection';
 import useExpandableGraph from '../hooks/useExpandableGraph';
 import useRecentChanges from '../hooks/useRecentChanges';
 import { getRootNodes } from './entityGraphShape';
+import { ASSIGNMENT_TYPE_STYLES } from '../utils/accessPackageStyles';
 
 const HEADER_FIELDS = ['catalogName', 'catalogId', 'description'];
 const HIDDEN_FIELDS = new Set([
   'displayName', ...HEADER_FIELDS, 'ValidFrom', 'ValidTo', 'extendedAttributes',
 ]);
-
-const ASSIGNMENT_TYPE_STYLES = {
-  'Auto-assigned': 'bg-green-100 dark:bg-green-900/30 text-green-800 dark:text-green-300 border-green-200 dark:border-green-700',
-  'Request-based': 'bg-blue-100 dark:bg-blue-900/30 text-blue-800 dark:text-blue-300 border-blue-200 dark:border-blue-700',
-  'Request-based with auto-removal': 'bg-orange-100 dark:bg-orange-900/30 text-orange-800 dark:text-orange-300 border-orange-200 dark:border-orange-700',
-  'Both': 'bg-purple-100 dark:bg-purple-900/30 text-purple-800 dark:text-purple-300 border-purple-200 dark:border-purple-700',
-};
 
 export default function AccessPackageDetailPage({ accessPackageId, cachedData, onCacheData, onClose, onOpenDetail }) {
   const { authFetch } = useAuth();
