+++ app/ui/src/utils/formatters.js
@@ -37,6 +37,55 @@ export function computeHistoryDiffs(history) {
   return diffs;
 }
 
+export function formatDateOnly(dateStr) {
+  if (!dateStr) return '';
+  const d = new Date(dateStr);
+  return d.toLocaleDateString(undefined, { day: 'numeric', month: 'short', year: 'numeric' });
+}
+
+export function formatDurationSeconds(seconds) {
+  if (seconds == null || isNaN(seconds)) return '—';
+  if (seconds < 60) return `${seconds}s`;
+  const h = Math.floor(seconds / 3600);
+  const m = Math.floor((seconds % 3600) / 60);
+  const s = seconds % 60;
+  if (h > 0) return s > 0 ? `${h}h ${m}m ${s}s` : `${h}h ${m}m`;
+  return s > 0 ? `${m}m ${s}s` : `${m}m`;
+}
+
+export function formatDurationMs(ms) {
+  if (ms == null || isNaN(ms)) return '—';
+  if (ms < 1000) return `${ms}ms`;
+  const s = Math.round(ms / 1000);
+  if (s < 60) return `${s}s`;
+  const m = Math.floor(s / 60);
+  return `${m}m ${s % 60}s`;
+}
+
+export function formatRelativeTime(isoStr) {
+  if (!isoStr) return 'never';
+  const diff = Date.now() - new Date(isoStr).getTime();
+  if (diff < 0) return 'in the future';
+  const minutes = Math.floor(diff / 60_000);
+  if (minutes < 1)  return 'just now';
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24)   return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  if (days < 30)    return `${days}d ago`;
+  const months = Math.floor(days / 30);
+  if (months < 12)  return `${months}mo ago`;
+  return `${Math.floor(months / 12)}y ago`;
+}
+
+export function formatCompactNumber(n) {
+  if (n == null) return '—';
+  if (n >= 1_000_000) return (n / 1_000_000).toFixed(1) + 'M';
+  if (n >= 10_000)    return (n / 1_000).toFixed(0) + 'k';
+  if (n >= 1_000)     return (n / 1_000).toFixed(1) + 'k';
+  return String(n);
+}
+
 export function friendlyLabel(key) {
   if (key === 'id') return 'GUID';
   return key.replace(/([A-Z])/g, ' $1').replace(/^./, s => s.toUpperCase()).trim();
